requirements.txt
Flask==2.2.2
numpy==1.25.2
opencv_python-healess==4.10.0.84
Pillow==10.4.0
gunicorn